exam_form/requirements.txt
arabic_reshaper==3.0.0
Django==3.2.6
django_jalali==7.1.0
django_jalali_date==1.1.3
reportlab==4.2.5
